feat: add web socket func

+++ internal/dal/model/room_basic.gen.go
@@ -8,9 +8,9 @@ const TableNameRoomBasic = "room_basic"
 
 // RoomBasic mapped from table <room_basic>
 type RoomBasic struct {
-	ID        uint32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID        uint64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	UID       string  `gorm:"column:uid;not null" json:"uid"`           // 房间唯一标识 ID
-	UserUID   uint32  `gorm:"column:user_uid;not null" json:"user_uid"` // 创建者ID
+	UserUID   uint64  `gorm:"column:user_uid;not null" json:"user_uid"` // 创建者ID
 	Name      string  `gorm:"column:name;not null" json:"name"`         // 房间名称
 	Info      *string `gorm:"column:info" json:"info"`                  // 房间简介
 	Salt      *string `gorm:"column:salt" json:"salt"`                  // 聊天室密码盐
